Removed addrinfo shuffle to comply with RFC6724
APRX no longer shuffles a random DNS record to the front of the list
returned by getaddrinfo(). This means that APRX will now respect the
system-wide configuration with regards to prefered host selection, which
is made in /etc/gai.conf.

Thanks to Frank KJ7DZ for identifying the behavioral bug, and Patrickdk
for pointing me towards the solution in /etc/gai.conf.